Extract PaneRenderState and related types to state.rs (refs #88)
Move out of callback.rs into a new state.rs module:
- PhysRect, hash_highlight_ranges
- CachedGlyph
- PaneRenderState + impl (new, is_content_dirty, is_appearance_dirty,
  update_fingerprint — all bumped to pub(crate))
- ImageTextureEntry
- ShapedGlyphEntry, TextRun, ShapeCacheKey

callback.rs drops from 1,185 → 971 lines. The remaining callback.rs is
now focused on the CallbackTrait impl (paint loop), TerminalGpuResources,
TerminalPaintCallback, and emit_custom_glyph.

No behavior changes.
